chore: move method into transformer
