+++ packages/discord.js/src/util/Transformers.js
@@ -16,4 +16,20 @@ function toSnakeCase(obj) {
   return Object.fromEntries(Object.entries(obj).map(([key, value]) => [snakeCase(key), toSnakeCase(value)]));
 }
 
-module.exports = { toSnakeCase };
+/**
+ * Transforms an API auto moderation action object to a camel-cased variant.
+ * @param {APIAutoModerationAction} autoModerationAction The action to transform
+ * @returns {AutoModerationAction}
+ * @ignore
+ */
+function _transformAPIAutoModerationAction(autoModerationAction) {
+  return {
+    type: autoModerationAction.type,
+    metadata: {
+      durationSeconds: autoModerationAction.metadata.duration_seconds ?? null,
+      channelId: autoModerationAction.metadata.channel_id ?? null,
+    },
+  };
+}
+
+module.exports = { toSnakeCase, _transformAPIAutoModerationAction };
